Done for today. Estimate block updated. test_3() worked. Now we have to do more rigorous testing to validate changes.

+++ trempy/estimate/clsEstimate.py
@@ -15,18 +15,25 @@ from trempy.shared.clsBase import BaseCls
 class EstimateClass(BaseCls):
     """This class manages all issues about the model estimation."""
 
-    def __init__(self, df, cutoffs, upper, marginals, questions, paras_obj, max_eval, version):
+    def __init__(self, df, cutoffs, questions, paras_obj, max_eval, version, **version_specific):
         """Init class."""
         self.attr = dict()
         self.attr['version'] = version
 
+        # Handle version-specific objects outside paras_obj.
+        if version in ['scaled_archimedean']:
+            # assert all(x in version_specific.keys() for x in ['marginals', 'upper'])
+            for key, value in version_specific.items():
+                self.attr[key] = value
+        elif version in ['nonstationary']:
+            # Currently nothing to do.
+            pass
+
         # Initialization attributes
         self.attr['paras_obj'] = paras_obj
         self.attr['questions'] = questions
-        self.attr['marginals'] = marginals
         self.attr['max_eval'] = max_eval
         self.attr['cutoffs'] = cutoffs
-        self.attr['upper'] = upper
         self.attr['df'] = df
 
         # Housekeeping attributes
@@ -51,21 +58,33 @@ class EstimateClass(BaseCls):
 
     def evaluate(self, x_optim_free_current):
         """Evaluate the criterion function during an estimation."""
-        # Distribute class attributes
+        # Distribute general class attributes
         paras_obj = self.attr['paras_obj']
         questions = self.attr['questions']
-        marginals = self.attr['marginals']
+        version = self.attr['version']
         cutoffs = self.attr['cutoffs']
-        upper = self.attr['upper']
         df = self.attr['df']
+
+        # Handle versions-specific objects outside para_obj
+        if version in ['scaled_archimedean']:
+            marginals = self.attr['marginals']
+            upper = self.attr['upper']
+            version_specific = {'upper': upper, 'marginals': marginals}
+        elif version in ['nonstationary']:
+            version_specific = dict()
 
         # Construct relevant set of parameters
         paras_obj.set_values('optim', 'free', x_optim_free_current)
         x_optim_all_current = paras_obj.get_values('optim', 'all')
         x_econ_all_current = paras_obj.get_values('econ', 'all')
 
+        # Get standard deviations. They have a larger index than nparas_econ.
+        nparas_econ = paras_obj.attr['nparas_econ']
+        stands = x_econ_all_current[nparas_econ:]
+
         # TODO: fix criterion function. I changed the parameters that are passed to it.
-        fval = criterion_function(df, questions, cutoffs, upper, marginals, *x_econ_all_current)
+        fval = criterion_function(df=df, questions=questions, cutoffs=cutoffs, paras_obj=paras_obj,
+                                  version=version, sds=stands, **version_specific)
 
         self._update_evaluation(fval, x_econ_all_current, x_optim_all_current)
 
@@ -74,12 +93,22 @@ class EstimateClass(BaseCls):
     def _update_evaluation(self, fval, x_econ_all_current, x_optim_all_current):
         """Update all attributes based on the new evaluation and write some information to file."""
         # Distribute class attribute
+
+        paras_obj = self.attr['paras_obj']
         questions = self.attr['questions']
-        marginals = self.attr['marginals']
-        upper = self.attr['upper']
+        version = self.attr['version']
+
+        if version in ['scaled_archimedean']:
+            marginals = self.attr['marginals']
+            upper = self.attr['upper']
+            version_specific = {'upper': upper, 'marginals': marginals}
+        elif version in ['nonstationary']:
+            version_specific = dict()
 
         # Update current information]
-        m_optimal = get_optimal_compensations(questions, upper, marginals, *x_econ_all_current[:5])
+        m_optimal = get_optimal_compensations(version=version, paras_obj=paras_obj,
+                                              questions=questions,
+                                              **version_specific)
 
         self.attr['x_econ_all_current'] = x_econ_all_current
         self.attr['m_optimal_current'] = m_optimal
@@ -137,6 +166,7 @@ class EstimateClass(BaseCls):
         # Distribute attributes
         para_labels = self.attr['paras_label']
         questions = self.attr['questions']
+        version = self.attr['version']
 
         # Update class attributes
         with open('est.trempy.info', 'w') as outfile:
@@ -149,17 +179,24 @@ class EstimateClass(BaseCls):
             line = ['', ''] + char_floats(args) + ['']
             outfile.write(fmt_.format(*line) + '\n\n')
 
+            # Economic Parameters
             outfile.write('\n {:<25}\n\n'.format('Economic Parameters'))
             line = ['Identifier', 'Label', 'Start', 'Step', 'Current']
             outfile.write(fmt_.format(*line) + '\n\n')
-            for i, _ in enumerate(range(len(questions) + 5)):
+            # Handle version
+            for i, _ in enumerate(range(len(questions) + len(PREFERENCE_PARAMETERS[version]))):
                 line = [i]
                 line += [para_labels[i]]
-                line += char_floats(self.attr['x_econ_all_start'][i])
-                line += char_floats(self.attr['x_econ_all_step'][i])
-                line += char_floats(self.attr['x_econ_all_current'][i])
-                outfile.write(fmt_.format(*line) + '\n')
+                # Handle optional arguments indicated by None value.
+                if self.attr['x_econ_all_start'][i] is None:
+                    continue
+                else:
+                    line += char_floats(self.attr['x_econ_all_start'][i])
+                    line += char_floats(self.attr['x_econ_all_step'][i])
+                    line += char_floats(self.attr['x_econ_all_current'][i])
+                    outfile.write(fmt_.format(*line) + '\n')
 
+            # Optimal Compensation
             outfile.write('\n\n {:<25}\n\n'.format('Optimal Compensations'))
             line = ['Questions', '', 'Start', 'Step', 'Current']
             outfile.write(fmt_.format(*line) + '\n\n')
@@ -170,6 +207,7 @@ class EstimateClass(BaseCls):
                 line += char_floats(self.attr['m_optimal_current'][q])
                 outfile.write(fmt_.format(*line) + '\n')
 
+            # Steps and Duration
             outfile.write('\n')
             fmt_ = '\n {:<25}   {:>25}\n'
             outfile.write(fmt_.format(*['Number of Evaluations', self.attr['num_eval']]))
@@ -188,19 +226,23 @@ class EstimateClass(BaseCls):
             line = ['Identifier', 'Label', 'Economic', 'Optimizer']
             outfile.write(fmt_.format(*line) + '\n\n')
 
-            for i, _ in enumerate(range(len(questions) + 5)):
+            for i, _ in enumerate(range(len(questions) + len(PREFERENCE_PARAMETERS[version]))):
                 line = [i]
                 line += [para_labels[i]]
+                # Handle optional arguments with None value
+                if (x_econ_all_current[i] or x_optim_all_current[i]) is None:
+                    continue
                 line += char_floats([x_econ_all_current[i], x_optim_all_current[i]])
                 outfile.write(fmt_.format(*line) + '\n')
             # We need to keep track of captured warnings.
             logger_obj.flush(outfile)
 
-            # We also record the results from the fitting of the copula.
-            outfile.write('\n')
-            with open('fit.copulpy.info') as infile:
-                outfile.write(infile.read())
-            os.remove('fit.copulpy.info')
+            if version in ['scaled_archimedean']:
+                # We also record the results from the fitting of the copula.
+                outfile.write('\n')
+                with open('fit.copulpy.info') as infile:
+                    outfile.write(infile.read())
+                os.remove('fit.copulpy.info')
 
         # We can determine the estimation if the number of requested function evaluations is
         # reached or the user requests a stop.
